requirements.txt
pycryptodome==3.19.1
pywin32==308
colorama==0.4.6 